医療的ケアの要否テーブル追加

+++ database/migrations/2023_11_08_125509_create_people_users_table.php
@@ -13,14 +13,17 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('people_users', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('person_id');
-            $table->unsignedBigInteger('user_id');
-            $table->foreign('person_id')->references('id')->on('people');
-            $table->foreign('user_id')->references('id')->on('users');
-            $table->timestamps();
-        });
+        // テーブルが存在しない場合にのみ作成
+        if (!Schema::hasTable('people_users')) {
+            Schema::create('people_users', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('person_id');
+                $table->unsignedBigInteger('user_id');
+                $table->foreign('person_id')->references('id')->on('people')->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -30,6 +33,9 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('people_users');
+        // テーブルが存在する場合にのみ削除
+        if (Schema::hasTable('people_users')) {
+            Schema::dropIfExists('people_users');
+        }
     }
 };
